refactor: Move bar-related view components into a dedicated /Bar directory and update routing.
