+++ src/views/Bar/Bar.vue
