+++ src/views/Bar/CocktailGlasses.vue
@@ -1,16 +1,16 @@
 <template>
   <div class="p-6 max-w-5xl mx-auto">
     <div class="flex justify-between items-center mb-6">
-      <h1 class="text-2xl font-bold">Украшения</h1>
+      <h1 class="text-2xl font-bold">Бокалы</h1>
       <a-button type="primary" @click="handleAdd">
         <template #icon><PlusOutlined /></template>
-        Добавить украшение
+        Добавить бокал
       </a-button>
     </div>
 
-    <!-- Список украшений -->
+    <!-- Список бокалов -->
     <a-row :gutter="[16, 16]">
-      <a-col v-for="item in decorations" :key="item._id" :xs="24" :sm="12" :md="8">
+      <a-col v-for="item in glasses" :key="item._id" :xs="24" :sm="12" :md="8">
         <a-card class="h-full flex flex-col" hoverable>
           <template #extra>
             <a-dropdown trigger="click">
@@ -24,10 +24,10 @@
                   </a-menu-item>
                   <a-menu-item danger>
                     <a-popconfirm
-                      title="Удалить украшение?"
+                      title="Удалить бокал?"
                       ok-text="Да"
                       cancel-text="Нет"
-                      @confirm="deleteDecoration(item._id)"
+                      @confirm="deleteGlass(item._id)"
                     >
                       <span><DeleteOutlined /> Удалить</span>
                     </a-popconfirm>
@@ -43,12 +43,12 @@
       </a-col>
     </a-row>
 
-    <AddDecorationDrawer
+    <AddGlassDrawer
       v-model:open="showAddDrawer"
       :width="drawerWidth"
-      :decorationToEdit="editingDecoration"
-      @decorationAdded="fetchDecorations"
-      @decorationUpdated="fetchDecorations"
+      :glassToEdit="editingGlass"
+      @glassAdded="fetchGlasses"
+      @glassUpdated="fetchGlasses"
     />
   </div>
 </template>
@@ -58,14 +58,14 @@ import { ref, onMounted, onUnmounted } from "vue";
 import axios from "axios";
 import { message } from "ant-design-vue";
 import { PlusOutlined, DeleteOutlined, MoreOutlined, EditOutlined } from "@ant-design/icons-vue";
-import { DECORATIONS_URL } from '../config/api.js';
-import { useAuthStore } from '../stores/auth';
-import AddDecorationDrawer from '../components/AddDecorationDrawer.vue';
+import { GLASSES_URL } from '../../config/api.js';
+import { useAuthStore } from '../../stores/auth';
+import AddGlassDrawer from '../../components/AddGlassDrawer.vue';
 
-const decorations = ref([]);
+const glasses = ref([]);
 const showAddDrawer = ref(false);
 const drawerWidth = ref('600px');
-const editingDecoration = ref(null);
+const editingGlass = ref(null);
 
 // Responsive drawer width
 const updateDrawerWidth = () => {
@@ -73,19 +73,19 @@ const updateDrawerWidth = () => {
 };
 
 const handleAdd = () => {
-  editingDecoration.value = null;
+  editingGlass.value = null;
   showAddDrawer.value = true;
 };
 
 const handleEdit = (item) => {
-  editingDecoration.value = item;
+  editingGlass.value = item;
   showAddDrawer.value = true;
 };
 
 onMounted(() => {
   updateDrawerWidth();
   window.addEventListener('resize', updateDrawerWidth);
-  fetchDecorations();
+  fetchGlasses();
 });
 
 onUnmounted(() => {
@@ -93,34 +93,34 @@ onUnmounted(() => {
 });
 
 // ===== Загрузка списка =====
-const fetchDecorations = async () => {
+const fetchGlasses = async () => {
   const authStore = useAuthStore();
   if (!authStore.selectedVenue) return;
 
   try {
-    const res = await axios.get(DECORATIONS_URL, {
+    const res = await axios.get(GLASSES_URL, {
       params: { venueId: authStore.selectedVenue._id },
       headers: { Authorization: `Bearer ${authStore.token}` }
     });
-    decorations.value = res.data || [];
+    glasses.value = res.data || [];
   } catch (e) {
     console.error(e);
-    message.error("Не удалось загрузить украшения");
+    message.error("Не удалось загрузить бокалы");
   }
 };
 
 // ===== Удаление =====
-const deleteDecoration = async (id) => {
+const deleteGlass = async (id) => {
   const authStore = useAuthStore();
   try {
-    await axios.delete(`${DECORATIONS_URL}/${id}`, {
+    await axios.delete(`${GLASSES_URL}/${id}`, {
       headers: { Authorization: `Bearer ${authStore.token}` }
     });
-    message.success("Украшение удалено");
-    await fetchDecorations();
+    message.success("Бокал удалён");
+    await fetchGlasses();
   } catch (e) {
     console.error(e);
-    message.error("Ошибка при удалении украшения");
+    message.error("Ошибка при удалении бокала");
   }
 };
 </script>
